requirements.txt
# Core LangChain
langchain>=0.3.9
langchain-core>=0.3.21

# Essential Tools
duckduckgo-search>=6.3.4
requests>=2.32.3
pandas>=2.0.0
pydantic>=2.0.0

# Web Frameworks & UI
# fasthtml>=0.6.9  # Commented out: not available on PyPI
monsterui>=0.0.8
fastlite>=0.0.6
streamlit>=1.28.0
uvicorn>=0.32.0
python-multipart>=0.0.6
fastapi>=0.109.0
starlette>=0.36.0

# Environment
python-dotenv>=1.0.1

# Additional packages
langchain-openai
langchain-anthropic
langchain-mistralai
langchain-groq
langgraph
beautifulsoup4
arxiv
lxml
markdown

# GraphRAG and Neo4j Integration
neo4j>=5.0.0
neo4j-graphrag>=0.1.0
sentence-transformers>=2.2.2
numpy>=1.21.0
scikit-learn>=1.0.0
networkx>=3.0
py2neo>=2021.2.3

# Vector embeddings and similarity
faiss-cpu>=1.7.4
chromadb>=0.4.0
openai>=1.0.0  # for embeddings

# Graph processing and analysis
igraph>=0.10.0
community>=1.0.0
python-louvain>=0.16 